svn merge -r 40086:40301 https://dart.googlecode.com/svn/branches/bleeding_edge trunk

git-svn-id: http://dart.googlecode.com/svn/trunk@40302 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/vm/scanner.h
@@ -26,7 +26,7 @@ class String;
 // GetStream() scans the entire source text and returns a stream of tokens.
 class Scanner : ValueObject {
  public:
-  typedef int32_t (*CharAtFunc)(const String& str, intptr_t index);
+  typedef uint16_t (*CharAtFunc)(const String& str, intptr_t index);
 
   // SourcePosition describes a text location in user friendly
   // terms of line number and column.
